Kalmanfilter

+++ raspberrypi3b/src/main.cpp
@@ -202,8 +202,8 @@ int sendFunc (char *sendData, int sendLen)
 {
     sem_wait(&semProcessFrameCplt);
     bpsPointTypeDef *data = (bpsPointTypeDef*)sendData;
-    data->setpointCoordinate[BPS_X_AXIS] = x;
-    data->setpointCoordinate[BPS_Y_AXIS] = KF.predict(y);
+    data->setpointCoordinate[BPS_X_AXIS] = (int)x;
+    data->setpointCoordinate[BPS_Y_AXIS] = (int)KF.predict(y);
     sem_post(&semSendDataCplt);
 }
